Simplify CTA resolution and impression tracking
- Move CTA name-lookup/unknown-id-warning logic from HtmlWriter into
  ConfigurationFile.ResolveCta, alongside the Ctas dictionary it
  resolves against, so any future caller gets the same fallback
  behavior for free instead of reimplementing it.
- Extract the CTA validation loop body into ValidateCta, matching
  the existing ValidateBranding convention in the same file.
- Reuse a single IntersectionObserver across htmx swaps instead of
  disconnecting and recreating it on every navigation; re-observing
  an already-observed element is a documented no-op.
- Reorder the sidebar CTA render condition so "show it" and "except
  in Codex" read as separate clauses (same truth table).
