+++ src/Elastic.Markdown/Layout/_TableOfContents.cshtml
@@ -72,7 +72,7 @@
 				</li>
 			}
 		</ul>
-		@if (Model.BuildType != BuildType.Codex && (Model.BuildType == BuildType.Assembler || Model.Cta.Name != Cta.DefaultName))
+		@if (Model.BuildType == BuildType.Assembler || (Model.Cta.Name != Cta.DefaultName && Model.BuildType != BuildType.Codex))
 		{
 		<div class="mt-4 hidden lg:block">
 			<div class="sidebar-trial-card">
